+++ src/Repositories/TempoRepository.php
@@ -2,6 +2,8 @@
 
 namespace App\Repositories;
 
+use App\Entities\Tempo\Attribute;
+
 /**
  * Class TempoRepository
  *
@@ -18,5 +20,17 @@ class TempoRepository extends Repository
     public function saveAttributes(string $attributes, $group = 'default')
     {
         $this->db->saveSetting('tempo:attributes:' . $group, $attributes);
+    }
+
+    /**
+     * Lists all saved tempo attributes
+     *
+     * @return Attribute[]
+     */
+    public function listAttributes()
+    {
+        return $this->db->all('settings', [
+            'key' => ['LIKE', 'tempo:attributes:%']
+        ], Attribute::class);
     }
 }